requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.2.10
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
cli-helpers==2.1.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
coverage==5.3
cryptography==3.2.1
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.1
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap-form==3.4
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-chartjs==2.2.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.7
django-filter==2.3.0
django-tables2==2.3.1
djangorestframework==3.11.1
dnspython==1.16.0
drf-writable-nested==0.6.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flake8==3.7.9
future==0.18.2
gitchangelog==3.0.4
gpg===1.13.1-unknown
green==3.2.2
gviz-api==1.9.0
html5lib==1.0.1
httpie==2.2.0
httplib2==0.14.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kaptan==0.5.10
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libtmux==0.8.2
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
mysqlclient==2.0.1
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.1.1
paramiko==2.6.0
parso==0.7.1
pdfarranger==1.4.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==1.10.3+dfsg
Pillow==7.2.0
pipenv==2020.8.13
prompt-toolkit==3.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-telegram-bot==13.0
pytz==2020.4
pytzdata==2020.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
ranger-fm==1.9.3
reportlab==3.5.49
requests==2.24.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
systemd-python==234
tablib==2.0.0
tabulate==0.8.6
terminaltables==3.1.0
tmuxp==1.5.4
tornado==6.1
traitlets==5.0.0
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.25.10
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2020.3.24
